+++ examples/ws2812-spi.rs
@@ -22,7 +22,7 @@ fn main() -> ! {
     let gpioa = dp.GPIOA.split();
 
     let spi = dp.SPI1.spi(
-        (gpioa.pa5, NoPin, gpioa.pa7),
+        (gpioa.pa5, NoPin::new(), gpioa.pa7),
         ws2812::MODE,
         3500.kHz(),
         &clocks,