+++ src/can.rs
@@ -109,7 +109,7 @@ impl<CAN: Instance> Can<CAN> {
     where
         NoPin: Into<CAN::Rx>,
     {
-        Self::new(usart, (tx_pin, NoPin))
+        Self::new(usart, (tx_pin, NoPin::new()))
     }
 }
 
@@ -118,7 +118,7 @@ impl<CAN: Instance> Can<CAN> {
     where
         NoPin: Into<CAN::Tx>,
     {
-        Self::new(usart, (NoPin, rx_pin))
+        Self::new(usart, (NoPin::new(), rx_pin))
     }
 }