+++ src/gpio/alt.rs
@@ -1,30 +1,30 @@
-use super::{Alternate, NoPin, OpenDrain, PinMode};
+use super::{Alternate, NoPin, OpenDrain, PinMode, PushPull};
 use crate::gpio::{self, Edge, ExtiPin};
 use crate::pac::EXTI;
 use crate::syscfg::SysCfg;
 
 macro_rules! pin {
-    ( $($(#[$docs:meta])* <$name:ident> for $(no: $NoPin:ty,)? [$(
-            $(#[$attr:meta])* $PX:ident<$A:literal $(, $Otype:ident)?>,
+    ( $($(#[$docs:meta])* <$name:ident> for $DefaultOtype:ident: $(no: $NoPin:ident,)? [$(
+            $(#[$attr:meta])* $PX:ident<$A:literal>,
         )*],)*) => {
         $(
             #[derive(Debug)]
             $(#[$docs])*
-            pub enum $name {
+            pub enum $name<Otype = $DefaultOtype> {
                 $(
-                    None($NoPin),
+                    None($NoPin<Otype>),
                 )?
 
                 $(
                     $(#[$attr])*
-                    $PX(gpio::$PX<Alternate<$A $(, $Otype)?>>),
+                    $PX(gpio::$PX<Alternate<$A, Otype>>),
                 )*
             }
 
-            impl crate::Sealed for $name { }
+            impl<Otype> crate::Sealed for $name<Otype> { }
 
             #[allow(unreachable_patterns)]
-            impl $name {
+            impl<Otype> $name<Otype> {
                 pub fn is_high(&self) -> bool {
                     !self.is_low()
                 }
@@ -39,7 +39,7 @@ macro_rules! pin {
                 }
             }
             #[allow(unreachable_patterns)]
-            impl ExtiPin for $name {
+            impl<Otype> ExtiPin for $name<Otype> {
                 fn make_interrupt_source(&mut self, _syscfg: &mut SysCfg) {
                     match self {
                         $(
@@ -100,19 +100,19 @@ macro_rules! pin {
             }
 
             $(
-                impl From<$NoPin> for $name {
-                    fn from(p: $NoPin) -> Self {
-                        Self::None(p)
+                impl<Otype> From<$NoPin> for $name<Otype> {
+                    fn from(_p: $NoPin) -> Self {
+                        Self::None($NoPin::new())
                     }
                 }
 
                 #[allow(irrefutable_let_patterns)]
-                impl TryFrom<$name> for $NoPin {
+                impl<Otype> TryFrom<$name<Otype>> for $NoPin {
                     type Error = ();
 
-                    fn try_from(a: $name) -> Result<Self, Self::Error> {
-                        if let $name::None(p) = a {
-                            Ok(p)
+                    fn try_from(a: $name<Otype>) -> Result<Self, Self::Error> {
+                        if let $name::None(_p) = a {
+                            Ok(NoPin::new())
                         } else {
                             Err(())
                         }
@@ -122,25 +122,32 @@ macro_rules! pin {
 
             $(
                 $(#[$attr])*
-                impl From<gpio::$PX> for $name {
+                impl<Otype> From<gpio::$PX> for $name<Otype>
+                where
+                    Alternate<$A, Otype>: PinMode,
+                {
                     fn from(p: gpio::$PX) -> Self {
                         Self::$PX(p.into_mode())
                     }
                 }
 
                 $(#[$attr])*
-                impl From<gpio::$PX<Alternate<$A $(, $Otype)?>>> for $name {
-                    fn from(p: gpio::$PX<Alternate<$A $(, $Otype)?>>) -> Self {
+                impl<Otype> From<gpio::$PX<Alternate<$A, Otype>>> for $name<Otype> {
+                    fn from(p: gpio::$PX<Alternate<$A, Otype>>) -> Self {
                         Self::$PX(p)
                     }
                 }
 
                 $(#[$attr])*
                 #[allow(irrefutable_let_patterns)]
-                impl<MODE: PinMode> TryFrom<$name> for gpio::$PX<MODE> {
+                impl<MODE, Otype> TryFrom<$name<Otype>> for gpio::$PX<MODE>
+                where
+                    MODE: PinMode,
+                    Alternate<$A, Otype>: PinMode,
+                {
                     type Error = ();
 
-                    fn try_from(a: $name) -> Result<Self, Self::Error> {
+                    fn try_from(a: $name<Otype>) -> Result<Self, Self::Error> {
                         if let $name::$PX(p) = a {
                             Ok(p.into_mode())
                         } else {
@@ -160,7 +167,7 @@ pub mod can1 {
     use super::*;
 
     pin! {
-        <Tx> for no:NoPin, [
+        <Tx> for PushPull: no:NoPin, [
             PA12<9>,
             PD1<9>,
 
@@ -186,7 +193,7 @@ pub mod can1 {
             PH13<9>,
         ],
 
-        <Rx> for no:NoPin, [
+        <Rx> for PushPull: no:NoPin, [
             PA11<9>,
             PD0<9>,
 
@@ -219,7 +226,7 @@ pub mod can2 {
     use super::*;
 
     pin! {
-        <Tx> for no:NoPin, [
+        <Tx> for PushPull: no:NoPin, [
             PB13<9>,
             PB6<9>,
 
@@ -227,7 +234,7 @@ pub mod can2 {
             PG12<9>,
         ],
 
-        <Rx> for no:NoPin, [
+        <Rx> for PushPull: no:NoPin, [
             PB12<9>,
             PB5<9>,
 
@@ -242,8 +249,8 @@ pub mod can3 {
     use super::*;
 
     pin! {
-        <Tx> for no:NoPin, [PA15<11>, PB4<11>,],
-        <Rx> for no:NoPin, [PA8<11>, PB3<11>,],
+        <Tx> for PushPull: no:NoPin, [PA15<11>, PB4<11>,],
+        <Rx> for PushPull: no:NoPin, [PA8<11>, PB3<11>,],
     }
 }
 
@@ -253,9 +260,9 @@ pub mod i2c1 {
     use super::*;
 
     pin! {
-        <Scl> for [PB6<4, OpenDrain>, PB8<4, OpenDrain>,],
+        <Scl> for OpenDrain: [PB6<4>, PB8<4>,],
 
-        <Sda> for [PB7<4, OpenDrain>, PB9<4, OpenDrain>,],
+        <Sda> for OpenDrain: [PB7<4>, PB9<4>,],
     }
 }
 
@@ -263,9 +270,9 @@ pub mod i2c2 {
     use super::*;
 
     pin! {
-        <Sda> for [
+        <Sda> for OpenDrain: [
             #[cfg(feature = "gpio-f446")]
-            PB3<4, OpenDrain>,
+            PB3<4>,
 
             #[cfg(any(
                 feature = "gpio-f401",
@@ -274,7 +281,7 @@ pub mod i2c2 {
                 feature = "gpio-f412",
                 feature = "gpio-f413",
             ))]
-            PB3<9, OpenDrain>,
+            PB3<9>,
 
             #[cfg(any(
                 feature = "gpio-f410",
@@ -282,7 +289,7 @@ pub mod i2c2 {
                 feature = "gpio-f412",
                 feature = "gpio-f413",
             ))]
-            PB9<9, OpenDrain>,
+            PB9<9>,
 
             #[cfg(any(
                 feature = "gpio-f417",
@@ -294,10 +301,10 @@ pub mod i2c2 {
                 feature = "gpio-f446",
                 feature = "gpio-f469",
             ))]
-            PB11<4, OpenDrain>,
+            PB11<4>,
 
             #[cfg(feature = "gpio-f446")]
-            PC12<4, OpenDrain>,
+            PC12<4>,
 
             #[cfg(any(
                 feature = "gpio-f417",
@@ -307,18 +314,18 @@ pub mod i2c2 {
                 feature = "gpio-f446",
                 feature = "gpio-f469",
             ))]
-            PF0<4, OpenDrain>,
+            PF0<4>,
 
             #[cfg(any(
                 feature = "gpio-f417",
                 feature = "gpio-f427",
                 feature = "gpio-f469",
             ))]
-            PH5<4, OpenDrain>,
+            PH5<4>,
         ],
 
-        <Scl> for [
-            PB10<4, OpenDrain>,
+        <Scl> for PushPull: [
+            PB10<4>,
 
             #[cfg(any(
                 feature = "gpio-f417",
@@ -328,14 +335,14 @@ pub mod i2c2 {
                 feature = "gpio-f446",
                 feature = "gpio-f469",
             ))]
-            PF1<4, OpenDrain>,
+            PF1<4>,
 
             #[cfg(any(
                 feature = "gpio-f417",
                 feature = "gpio-f427",
                 feature = "gpio-f469",
             ))]
-            PH4<4, OpenDrain>,
+            PH4<4>,
         ],
     }
 }
@@ -345,22 +352,22 @@ pub mod i2c3 {
     use super::*;
 
     pin! {
-        <Scl> for [
-            PA8<4, OpenDrain>,
+        <Scl> for OpenDrain: [
+            PA8<4>,
 
             #[cfg(any(
                 feature = "gpio-f417",
                 feature = "gpio-f427",
                 feature = "gpio-f469",
             ))]
-            PH7<4, OpenDrain>,
+            PH7<4>,
         ],
 
-        <Sda> for [
-            PC9<4, OpenDrain>,
+        <Sda> for OpenDrain: [
+            PC9<4>,
 
             #[cfg(feature = "gpio-f446")]
-            PB4<4, OpenDrain>,
+            PB4<4>,
 
             #[cfg(any(
                 feature = "gpio-f401",
@@ -368,21 +375,21 @@ pub mod i2c3 {
                 feature = "gpio-f412",
                 feature = "gpio-f413",
             ))]
-            PB4<9, OpenDrain>,
+            PB4<9>,
 
             #[cfg(any(
                 feature = "gpio-f411",
                 feature = "gpio-f412",
                 feature = "gpio-f413",
             ))]
-            PB8<9, OpenDrain>,
+            PB8<9>,
 
             #[cfg(any(
                 feature = "gpio-f417",
                 feature = "gpio-f427",
                 feature = "gpio-f469",
             ))]
-            PH8<4, OpenDrain>,
+            PH8<4>,
         ],
     }
 }
@@ -391,21 +398,21 @@ pub mod i2c3 {
 pub mod fmpi2c1 {
     use super::*;
     pin! {
-        <Sda> for [
-            PB3<4, OpenDrain>,
-            PB14<4, OpenDrain>,
-            PC7<4, OpenDrain>,
-            PD13<4, OpenDrain>,
-            PD15<4, OpenDrain>,
-            PF15<4, OpenDrain>,
+        <Sda> for OpenDrain: [
+            PB3<4>,
+            PB14<4>,
+            PC7<4>,
+            PD13<4>,
+            PD15<4>,
+            PF15<4>,
         ],
-        <Scl> for [
-            PB10<9, OpenDrain>,
-            PB15<4, OpenDrain>,
-            PC6<4, OpenDrain>,
-            PD12<4, OpenDrain>,
-            PD14<4, OpenDrain>,
-            PF14<4, OpenDrain>,
+        <Scl> for OpenDrain: [
+            PB10<9>,
+            PB15<4>,
+            PC6<4>,
+            PD12<4>,
+            PD14<4>,
+            PF14<4>,
         ],
     }
 }
@@ -415,15 +422,15 @@ pub mod fmpi2c1 {
 pub mod spi1 {
     use super::*;
     pin! {
-        <Sck> for no:NoPin, [PA5<5>, PB3<5>,],
+        <Sck> for PushPull: no:NoPin, [PA5<5>, PB3<5>,],
 
-        <Miso> for no:NoPin, [PA6<5>, PB4<5>,],
+        <Miso> for PushPull: no:NoPin, [PA6<5>, PB4<5>,],
 
-        <Mosi> for no:NoPin, [
+        <Mosi> for PushPull: no:NoPin, [
             PA7<5>, PB5<5>,
         ],
 
-        <Nss> for [
+        <Nss> for PushPull: no:NoPin, [
             #[cfg(any(
                 feature = "gpio-f410",
                 feature = "gpio-f411",
@@ -432,6 +439,7 @@ pub mod spi1 {
                 feature = "gpio-f446"
             ))]
             PA4<5>,
+
             #[cfg(any(
                 feature = "gpio-f410",
                 feature = "gpio-f411",
@@ -447,7 +455,7 @@ pub mod spi1 {
 pub mod spi2 {
     use super::*;
     pin! {
-        <Sck> for no:NoPin, [
+        <Sck> for PushPull: no:NoPin, [
             PB10<5>,
             PB13<5>,
 
@@ -486,7 +494,7 @@ pub mod spi2 {
             PI1<5>,
         ],
 
-        <Miso> for no:NoPin, [
+        <Miso> for PushPull: no:NoPin, [
             PB14<5>, PC2<5>,
 
             #[cfg(any(
@@ -500,7 +508,7 @@ pub mod spi2 {
             PA12<5>,
         ],
 
-        <Mosi> for no:NoPin, [
+        <Mosi> for PushPull: no:NoPin, [
             PB15<5>,
             PC3<5>,
 
@@ -521,7 +529,7 @@ pub mod spi2 {
             PC1<7>,
         ],
 
-        <Nss> for [
+        <Nss> for PushPull: no:NoPin, [
             PB9<5>,
             PB12<5>,
 
@@ -548,7 +556,7 @@ pub mod spi2 {
 pub mod spi3 {
     use super::*;
     pin! {
-        <Sck> for no:NoPin, [
+        <Sck> for PushPull: no:NoPin, [
             PB3<6>,
             PC10<6>,
 
@@ -560,9 +568,9 @@ pub mod spi3 {
             PB12<7>,
         ],
 
-        <Miso> for no:NoPin, [PB4<6>, PC11<6>,],
+        <Miso> for PushPull: no:NoPin, [PB4<6>, PC11<6>,],
 
-        <Mosi> for no:NoPin, [
+        <Mosi> for PushPull: no:NoPin, [
             PB5<6>,
             PC12<6>,
 
@@ -587,7 +595,7 @@ pub mod spi3 {
             PD0<6>,
         ],
 
-        <Nss> for [PA4<6>, PA15<6>,],
+        <Nss> for PushPull: [PA4<6>, PA15<6>,],
     }
 }
 
@@ -596,7 +604,7 @@ pub mod spi4 {
     use super::*;
 
     pin! {
-        <Sck> for no:NoPin, [
+        <Sck> for PushPull: no:NoPin, [
             PE2<5>,
             PE12<5>,
 
@@ -611,7 +619,7 @@ pub mod spi4 {
             PG11<6>,
         ],
 
-        <Miso> for no:NoPin, [
+        <Miso> for PushPull: no:NoPin, [
             PE5<5>,
             PE13<5>,
 
@@ -629,7 +637,7 @@ pub mod spi4 {
             PD0<5>,
         ],
 
-        <Mosi> for no:NoPin, [
+        <Mosi> for PushPull: no:NoPin, [
             PE6<5>,
             PE14<5>,
 
@@ -644,7 +652,7 @@ pub mod spi4 {
             PG13<6>,
         ],
 
-        <Nss> for [
+        <Nss> for PushPull: no:NoPin, [
             #[cfg(any(
                 feature = "gpio-f411",
                 feature = "gpio-f412",
@@ -674,7 +682,7 @@ pub mod spi5 {
     use super::*;
 
     pin! {
-        <Sck> for no:NoPin, [
+        <Sck> for PushPull: no:NoPin, [
             #[cfg(any(
                 feature = "gpio-f410",
                 feature = "gpio-f411",
@@ -709,7 +717,7 @@ pub mod spi5 {
             PH6<5>,
         ],
 
-        <Miso> for no:NoPin, [
+        <Miso> for PushPull: no:NoPin, [
             #[cfg(any(
                 feature = "gpio-f410",
                 feature = "gpio-f411",
@@ -745,7 +753,7 @@ pub mod spi5 {
             PH7<5>,
         ],
 
-        <Mosi> for no:NoPin, [
+        <Mosi> for PushPull: no:NoPin, [
             #[cfg(any(
                 feature = "gpio-f410",
                 feature = "gpio-f411",
@@ -789,7 +797,7 @@ pub mod spi5 {
             PF11<5>,
         ],
 
-        <Nss> for [
+        <Nss> for PushPull: no:NoPin, [
             #[cfg(any(
                 feature = "gpio-f410",
                 feature = "gpio-f411",
@@ -820,28 +828,28 @@ pub mod spi6 {
     use super::*;
 
     pin! {
-        <Sck> for no:NoPin, [PG13<5>,],
-        <Miso> for no:NoPin, [PG12<5>,],
-        <Mosi> for no:NoPin, [PG14<5>,],
-        <Nss> for [],
+        <Sck> for PushPull: no:NoPin, [PG13<5>,],
+        <Miso> for PushPull: no:NoPin, [PG12<5>,],
+        <Mosi> for PushPull: no:NoPin, [PG14<5>,],
+        <Nss> for PushPull: no:NoPin, [],
     }
 }
 
-// SPI pins for I2S mode
+// SPI pins for PushPull: I2S mode
 pub mod i2s1 {
     use super::*;
 
     pin! {
-        <Ck> for [
+        <Ck> for PushPull: [
             PA5<5>,
             PB3<5>,
         ],
-        <Sd> for [
+        <Sd> for PushPull: [
             PA7<5>,
             PB5<5>,
         ],
 
-        <Ws> for [
+        <Ws> for PushPull: no:NoPin, [
             #[cfg(any(
                 feature = "gpio-f410",
                 feature = "gpio-f411",
@@ -861,7 +869,7 @@ pub mod i2s1 {
             PA15<5>,
         ],
 
-        <Mck> for no:NoPin, [
+        <Mck> for PushPull: no:NoPin, [
             #[cfg(any(
                 feature = "gpio-f412",
                 feature = "gpio-f413",
@@ -882,7 +890,7 @@ pub mod i2s2 {
     use super::*;
 
     pin! {
-        <Ck> for [
+        <Ck> for PushPull: [
             PB10<5>,
             PB13<5>,
 
@@ -921,7 +929,7 @@ pub mod i2s2 {
             PA9<5>,
         ],
 
-        <Sd> for [
+        <Sd> for PushPull: [
             PB15<5>,
             PC3<5>,
 
@@ -942,7 +950,7 @@ pub mod i2s2 {
             PC1<5>,
         ],
 
-        <Ws> for [
+        <Ws> for PushPull: [
             PB9<5>,
             PB12<5>,
 
@@ -963,7 +971,7 @@ pub mod i2s2 {
             PD1<7>,
         ],
 
-        <Mck> for no:NoPin, [
+        <Mck> for PushPull: no:NoPin, [
             PC6<5>,
 
             #[cfg(any(
@@ -971,8 +979,8 @@ pub mod i2s2 {
                 feature = "gpio-f412",
                 feature = "gpio-f413",
             ))]
-            PA3<5>,
 
+            PA3<5>,
             #[cfg(any(
                 feature = "gpio-f411",
                 feature = "gpio-f412",
@@ -988,7 +996,7 @@ pub mod i2s3 {
     use super::*;
 
     pin! {
-        <Ck> for [
+        <Ck> for PushPull: [
             PB3<6>,
             PC10<6>,
 
@@ -1000,7 +1008,7 @@ pub mod i2s3 {
             PB12<7>,
         ],
 
-        <Sd> for [
+        <Sd> for PushPull: [
             PB5<6>,
             PC12<6>,
 
@@ -1025,12 +1033,12 @@ pub mod i2s3 {
             PD0<6>,
         ],
 
-        <Ws> for [
+        <Ws> for PushPull: [
             PA4<6>,
             PA15<6>,
         ],
 
-        <Mck> for no:NoPin, [
+        <Mck> for PushPull: no:NoPin, [
             PC7<6>,
 
             #[cfg(any(
@@ -1048,7 +1056,7 @@ pub mod i2s4 {
     use super::*;
 
     pin! {
-        <Ck> for [
+        <Ck> for PushPull: [
             PE2<5>,
             PE12<5>,
 
@@ -1063,7 +1071,7 @@ pub mod i2s4 {
             PG11<6>,
         ],
 
-        <Sd> for [
+        <Sd> for PushPull: [
             PE6<5>,
             PE14<5>,
 
@@ -1078,23 +1086,21 @@ pub mod i2s4 {
             PG13<6>,
         ],
 
-        <Mck> for no:NoPin, [ ],
+        <Mck> for PushPull: no:NoPin, [ ],
 
-        <Ws> for [
+        <Ws> for PushPull: no:NoPin, [
             #[cfg(any(
                 feature = "gpio-f411",
                 feature = "gpio-f412",
                 feature = "gpio-f413",
             ))]
             PB12<6>,
-
             #[cfg(any(
                 feature = "gpio-f411",
                 feature = "gpio-f412",
                 feature = "gpio-f413",
             ))]
             PE4<5>,
-
             #[cfg(any(
                 feature = "gpio-f411",
                 feature = "gpio-f412",
@@ -1110,7 +1116,7 @@ pub mod i2s5 {
     use super::*;
 
     pin! {
-        <Ck> for [
+        <Ck> for PushPull: [
             #[cfg(any(
                 feature = "gpio-f410",
                 feature = "gpio-f411",
@@ -1146,7 +1152,7 @@ pub mod i2s5 {
             PH6<5>,
         ],
 
-        <Sd> for [
+        <Sd> for PushPull: [
             #[cfg(any(
                 feature = "gpio-f410",
                 feature = "gpio-f411",
@@ -1190,9 +1196,9 @@ pub mod i2s5 {
             PF11<5>,
         ],
 
-        <Mck> for no:NoPin, [ ],
+        <Mck> for PushPull: no:NoPin, [ ],
 
-        <Ws> for [
+        <Ws> for PushPull: no:NoPin, [
             #[cfg(any(
                 feature = "gpio-f410",
                 feature = "gpio-f411",
@@ -1223,8 +1229,8 @@ pub mod i2s6 {
     use super::*;
 
     pin! {
-        <Ck> for [PG13<5>,],
-        <Sd> for [PG14<5>,],
+        <Ck> for PushPull: [PG13<5>,],
+        <Sd> for PushPull: [PG14<5>,],
     }
 }
 
@@ -1234,7 +1240,7 @@ pub mod usart1 {
     use super::*;
 
     pin! {
-        <Tx> for no:NoPin, [
+        <Tx> for PushPull: no:NoPin, [
             PA9<7>,
             PB6<7>,
 
@@ -1247,7 +1253,7 @@ pub mod usart1 {
             PA15<7>,
         ],
 
-        <Rx> for no:NoPin, [
+        <Rx> for PushPull: no:NoPin, [
             PA10<7>,
             PB7<7>,
 
@@ -1266,7 +1272,7 @@ pub mod usart2 {
     use super::*;
 
     pin! {
-        <Tx> for no:NoPin, [
+        <Tx> for PushPull: no:NoPin, [
             PA2<7>,
 
             #[cfg(any(
@@ -1282,7 +1288,7 @@ pub mod usart2 {
             PD5<7>,
         ],
 
-        <Rx> for no:NoPin, [
+        <Rx> for PushPull: no:NoPin, [
             PA3<7>,
 
             #[cfg(any(
@@ -1305,7 +1311,7 @@ pub mod usart3 {
     use super::*;
 
     pin! {
-        <Tx> for no:NoPin, [
+        <Tx> for PushPull: no:NoPin, [
             PB10<7>,
 
             #[cfg(any(
@@ -1329,7 +1335,7 @@ pub mod usart3 {
             PD8<7>,
         ],
 
-        <Rx> for no:NoPin, [
+        <Rx> for PushPull: no:NoPin, [
             PB11<7>,
 
             #[cfg(any(
@@ -1366,7 +1372,7 @@ pub mod usart6 {
     use super::*;
 
     pin! {
-        <Tx> for no:NoPin, [
+        <Tx> for PushPull: no:NoPin, [
             PC6<8>,
 
             #[cfg(any(
@@ -1381,7 +1387,7 @@ pub mod usart6 {
             #[cfg(feature = "gpiog")]
             PG14<8>,
         ],
-        <Rx> for no:NoPin, [
+        <Rx> for PushPull: no:NoPin, [
             PC7<8>,
 
             #[cfg(any(
@@ -1404,7 +1410,7 @@ pub mod uart4 {
     use super::*;
 
     pin! {
-        <Tx> for no:NoPin, [
+        <Tx> for PushPull: no:NoPin, [
             PA0<8>,
 
             #[cfg(any(
@@ -1425,7 +1431,7 @@ pub mod uart4 {
             PD10<8>,
         ],
 
-        <Rx> for no:NoPin, [
+        <Rx> for PushPull: no:NoPin, [
             PA1<8>,
 
             #[cfg(any(
@@ -1453,7 +1459,7 @@ pub mod uart5 {
     use super::*;
 
     pin! {
-        <Tx> for no:NoPin, [
+        <Tx> for PushPull: no:NoPin, [
             PC12<8>,
 
             #[cfg(feature = "gpio-f446")]
@@ -1469,7 +1475,7 @@ pub mod uart5 {
             PB13<11>,
         ],
 
-        <Rx> for no:NoPin, [
+        <Rx> for PushPull: no:NoPin, [
             PD2<8>,
 
             #[cfg(feature = "gpio-f446")]
@@ -1492,7 +1498,7 @@ pub mod uart7 {
     use super::*;
 
     pin! {
-        <Tx> for no:NoPin, [
+        <Tx> for PushPull: no:NoPin, [
             #[cfg(any(
                 feature = "gpio-f401",
                 feature = "gpio-f411",
@@ -1522,7 +1528,7 @@ pub mod uart7 {
             PB4<8>,
         ],
 
-        <Rx> for no:NoPin, [
+        <Rx> for PushPull: no:NoPin, [
             #[cfg(any(
                 feature = "gpio-f401",
                 feature = "gpio-f411",
@@ -1559,7 +1565,7 @@ pub mod uart8 {
     use super::*;
 
     pin! {
-        <Tx> for no:NoPin, [
+        <Tx> for PushPull: no:NoPin, [
             #[cfg(any(
                 feature = "gpio-f401",
                 feature = "gpio-f411",
@@ -1576,7 +1582,7 @@ pub mod uart8 {
             PF9<8>,
         ],
 
-        <Rx> for no:NoPin, [
+        <Rx> for PushPull: no:NoPin, [
             #[cfg(any(
                 feature = "gpio-f401",
                 feature = "gpio-f411",
@@ -1600,8 +1606,8 @@ pub mod uart9 {
     use super::*;
 
     pin! {
-        <Tx> for no:NoPin, [PD15<11>, PG1<11>,],
-        <Rx> for no:NoPin, [PD14<11>, PG0<11>,],
+        <Tx> for PushPull: no:NoPin, [PD15<11>, PG1<11>,],
+        <Rx> for PushPull: no:NoPin, [PD14<11>, PG0<11>,],
     }
 }
 
@@ -1610,8 +1616,8 @@ pub mod uart10 {
     use super::*;
 
     pin! {
-        <Tx> for no:NoPin, [PE3<11>, PG12<11>,],
-        <Rx> for no:NoPin, [PE2<11>, PG11<11>,],
+        <Tx> for PushPull: no:NoPin, [PE3<11>, PG12<11>,],
+        <Rx> for PushPull: no:NoPin, [PE2<11>, PG11<11>,],
     }
 }
 
@@ -1620,19 +1626,19 @@ pub mod sdio {
     use super::*;
 
     pin! {
-        <Clk> for [
+        <Clk> for PushPull: [
             PC12<12>,
 
             #[cfg(any(feature = "gpio-f412", feature = "gpio-f413", feature = "gpio-f411"))]
             PB15<12>,
         ],
-        <Cmd> for [
+        <Cmd> for PushPull: [
             PD2<12>,
 
             #[cfg(any(feature = "gpio-f412", feature = "gpio-f413", feature = "gpio-f411"))]
             PA6<12>,
         ],
-        <D0> for [
+        <D0> for PushPull: [
             PC8<12>,
 
             #[cfg(any(feature = "gpio-f412", feature = "gpio-f413", feature = "gpio-f411"))]
@@ -1644,37 +1650,37 @@ pub mod sdio {
             #[cfg(feature = "gpio-f411")]
             PB7<12>,
         ],
-        <D1> for [
+        <D1> for PushPull: [
             PC9<12>,
 
             #[cfg(any(feature = "gpio-f412", feature = "gpio-f413", feature = "gpio-f411"))]
             PA8<12>,
         ],
-        <D2> for [
+        <D2> for PushPull: [
             PC10<12>,
 
             #[cfg(any(feature = "gpio-f412", feature = "gpio-f413", feature = "gpio-f411"))]
             PA9<12>,
         ],
-        <D3> for [
+        <D3> for PushPull: [
             PC11<12>,
 
             #[cfg(any(feature = "gpio-f412", feature = "gpio-f413", feature = "gpio-f411"))]
             PB5<12>,
         ],
-        <D4> for [
+        <D4> for PushPull: [
             PB8<12>,
         ],
-        <D5> for [
+        <D5> for PushPull: [
             PB9<12>,
         ],
-        <D6> for [
+        <D6> for PushPull: [
             PC6<12>,
 
             #[cfg(any(feature = "gpio-f412", feature = "gpio-f413", feature = "gpio-f411"))]
             PB14<12>,
         ],
-        <D7> for [
+        <D7> for PushPull: [
             PC7<12>,
 
             #[cfg(any(feature = "gpio-f412", feature = "gpio-f413", feature = "gpio-f411"))]
@@ -1691,7 +1697,7 @@ pub mod fsmc {
     // All FSMC/FMC pins use 12
     pin! {
         /// A pin that can be used for data bus 0
-        <D0> for [
+        <D0> for PushPull: [
             PD14<12>,
 
             #[cfg(any(feature = "gpio-f412", feature = "gpio-f413"))]
@@ -1699,7 +1705,7 @@ pub mod fsmc {
         ],
 
         /// A pin that can be used for data bus 1
-        <D1> for [
+        <D1> for PushPull: [
             PD15<12>,
 
             #[cfg(any(feature = "gpio-f412", feature = "gpio-f413"))]
@@ -1707,7 +1713,7 @@ pub mod fsmc {
         ],
 
         /// A pin that can be used for data bus 2
-        <D2> for [
+        <D2> for PushPull: [
             PD0<12>,
 
             #[cfg(any(feature = "gpio-f412", feature = "gpio-f413"))]
@@ -1715,7 +1721,7 @@ pub mod fsmc {
         ],
 
         /// A pin that can be used for data bus 3
-        <D3> for [
+        <D3> for PushPull: [
             PD1<12>,
 
             #[cfg(any(feature = "gpio-f412", feature = "gpio-f413"))]
@@ -1723,7 +1729,7 @@ pub mod fsmc {
         ],
 
         /// A pin that can be used for data bus 4
-        <D4> for [
+        <D4> for PushPull: [
             PE7<12>,
 
             #[cfg(any(feature = "gpio-f412", feature = "gpio-f413"))]
@@ -1731,7 +1737,7 @@ pub mod fsmc {
         ],
 
         /// A pin that can be used for data bus 5
-        <D5> for [
+        <D5> for PushPull: [
             PE8<12>,
 
             #[cfg(any(feature = "gpio-f412", feature = "gpio-f413"))]
@@ -1739,7 +1745,7 @@ pub mod fsmc {
         ],
 
         /// A pin that can be used for data bus 6
-        <D6> for [
+        <D6> for PushPull: [
             PE9<12>,
 
             #[cfg(any(feature = "gpio-f412", feature = "gpio-f413"))]
@@ -1747,7 +1753,7 @@ pub mod fsmc {
         ],
 
         /// A pin that can be used for data bus 7
-        <D7> for [
+        <D7> for PushPull: [
             PE10<12>,
 
             #[cfg(any(feature = "gpio-f412", feature = "gpio-f413"))]
@@ -1755,32 +1761,32 @@ pub mod fsmc {
         ],
 
         /// A pin that can be used for data bus 8
-        <D8> for [
+        <D8> for PushPull: [
             PE11<12>,
         ],
 
         /// A pin that can be used for data bus 9
-        <D9> for [
+        <D9> for PushPull: [
             PE12<12>,
         ],
 
         /// A pin that can be used for data bus 10
-        <D10> for [
+        <D10> for PushPull: [
             PE13<12>,
         ],
 
         /// A pin that can be used for data bus 11
-        <D11> for [
+        <D11> for PushPull: [
             PE14<12>,
         ],
 
         /// A pin that can be used for data bus 12
-        <D12> for [
+        <D12> for PushPull: [
             PE15<12>,
         ],
 
         /// A pin that can be used for data bus 13
-        <D13> for [
+        <D13> for PushPull: [
             PD8<12>,
 
             #[cfg(any(feature = "gpio-f412", feature = "gpio-f413"))]
@@ -1788,17 +1794,17 @@ pub mod fsmc {
         ],
 
         /// A pin that can be used for data bus 14
-        <D14> for [
+        <D14> for PushPull: [
             PD9<12>,
         ],
 
         /// A pin that can be used for data bus 15
-        <D15> for [
+        <D15> for PushPull: [
             PD10<12>,
         ],
 
         /// A pin that can be used for the output enable (read enable, NOE) signal
-        <ReadEnable> for [
+        <ReadEnable> for PushPull: [
             PD4<12>,
 
             #[cfg(any(feature = "gpio-f412", feature = "gpio-f413"))]
@@ -1806,7 +1812,7 @@ pub mod fsmc {
         ],
 
         /// A pin that can be used for the write enable (NOE) signal
-        <WriteEnable> for [
+        <WriteEnable> for PushPull: [
             PD5<12>,
 
             #[cfg(any(feature = "gpio-f412", feature = "gpio-f413"))]
@@ -1819,7 +1825,7 @@ pub mod fsmc {
         /// A pin that can be used as one bit of the memory address
         ///
         /// This is used to switch between data and command mode.
-        <Address> for [
+        <Address> for PushPull: [
             PD11<12>,
             PD12<12>,
             PD13<12>,
@@ -1851,22 +1857,22 @@ pub mod fsmc {
         ],
 
         /// A pin that can be used to enable a memory device on sub-bank 1
-        <ChipSelect1> for [
+        <ChipSelect1> for PushPull: [
             PD7<12>,
         ],
 
         /// A pin that can be used to enable a memory device on sub-bank 2
-        <ChipSelect2> for [
+        <ChipSelect2> for PushPull: [
             PG9<12>,
         ],
 
         /// A pin that can be used to enable a memory device on sub-bank 3
-        <ChipSelect3> for [
+        <ChipSelect3> for PushPull: [
             PG10<12>,
         ],
 
         /// A pin that can be used to enable a memory device on sub-bank 4
-        <ChipSelect4> for [
+        <ChipSelect4> for PushPull: [
             PG12<12>,
 
             #[cfg(any(feature = "gpio-f412", feature = "gpio-f413"))]
@@ -1884,10 +1890,10 @@ pub mod otg_fs {
     use super::*;
 
     pin! {
-        <Dm> for [
+        <Dm> for PushPull: [
             PA11<10>,
         ],
-        <Dp> for [
+        <Dp> for PushPull: [
             PA12<10>,
         ],
     }
@@ -1898,10 +1904,10 @@ pub mod otg_hs {
     use super::*;
 
     pin! {
-        <Dm> for [
+        <Dm> for PushPull: [
             PB14<12>,
         ],
-        <Dp> for [
+        <Dp> for PushPull: [
             PB15<12>,
         ],
     }
@@ -1911,52 +1917,52 @@ pub mod tim1 {
     use super::*;
 
     pin! {
-        <C1> for [
+        <C1> for PushPull: no:NoPin, [
             PA8<1>,
 
             #[cfg(not(feature = "gpio-f410"))]
             PE9<1>,
         ],
-        <C2> for [
+        <C2> for PushPull: no:NoPin, [
             PA9<1>,
 
             #[cfg(not(feature = "gpio-f410"))]
             PE11<1>,
         ],
-        <C3> for [
+        <C3> for PushPull: no:NoPin, [
             PA10<1>,
 
             #[cfg(not(feature = "gpio-f410"))]
             PE13<1>,
         ],
-        <C4> for [
+        <C4> for PushPull: no:NoPin, [
             PA11<1>,
 
             #[cfg(not(feature = "gpio-f410"))]
             PE14<1>,
         ],
-        <NC1> for [
+        <NC1> for PushPull: no:NoPin, [
             PA7<1>,
             PB13<1>,
 
             #[cfg(not(feature = "gpio-f410"))]
             PE8<1>,
         ],
-        <NC2> for [
+        <NC2> for PushPull: no:NoPin, [
             PB0<1>,
             PB14<1>,
 
             #[cfg(not(feature = "gpio-f410"))]
             PE10<1>,
         ],
-        <NC3> for [
+        <NC3> for PushPull: no:NoPin, [
             PB1<1>,
             PB15<1>,
 
             #[cfg(not(feature = "gpio-f410"))]
             PE12<1>,
         ],
-        <Etr> for [
+        <Etr> for PushPull: no:NoPin, [
             PA12<1>,
 
             #[cfg(not(feature = "gpio-f410"))]
@@ -1965,7 +1971,7 @@ pub mod tim1 {
             #[cfg(any(feature = "gpio-f412", feature = "gpio-f413"))]
             PF10<1>,
         ],
-        <BkIn> for [
+        <BkIn> for PushPull: no:NoPin, [
             PA6<1>,
             PB12<1>,
 
@@ -1978,25 +1984,25 @@ pub mod tim1 {
     use crate::timer::ChannelPin;
 
     impl ChannelPin<0> for TIM1 {
-        type Pin = C1;
+        type Pin<Otype> = C1<Otype>;
     }
     impl ChannelPin<1> for TIM1 {
-        type Pin = C2;
+        type Pin<Otype> = C2<Otype>;
     }
     impl ChannelPin<2> for TIM1 {
-        type Pin = C3;
+        type Pin<Otype> = C3<Otype>;
     }
     impl ChannelPin<3> for TIM1 {
-        type Pin = C4;
+        type Pin<Otype> = C4<Otype>;
     }
     impl ChannelPin<0, true> for TIM1 {
-        type Pin = NC1;
+        type Pin<Otype> = NC1<Otype>;
     }
     impl ChannelPin<1, true> for TIM1 {
-        type Pin = NC2;
+        type Pin<Otype> = NC2<Otype>;
     }
     impl ChannelPin<2, true> for TIM1 {
-        type Pin = NC3;
+        type Pin<Otype> = NC3<Otype>;
     }
 }
 
@@ -2005,7 +2011,7 @@ pub mod tim2 {
     use super::*;
 
     pin! {
-        <C1> for [
+        <C1> for PushPull: no:NoPin, [
             PA0<1>,
             PA5<1>,
             PA15<1>,
@@ -2013,25 +2019,25 @@ pub mod tim2 {
             #[cfg(feature = "gpio-f446")]
             PB8<1>,
         ],
-        <C2> for [
+        <C2> for PushPull: no:NoPin, [
             PA1<1>,
             PB3<1>,
 
             #[cfg(feature = "gpio-f446")]
             PB9<1>,
         ],
-        <C3> for [
+        <C3> for PushPull: no:NoPin, [
             PA2<1>,
             PB10<1>,
         ],
-        <C4> for [
+        <C4> for PushPull: no:NoPin, [
             PA3<1>,
             PB11<1>,
 
             #[cfg(feature = "gpio-f446")]
             PB2<1>,
         ],
-        <Etr> for [
+        <Etr> for PushPull: no:NoPin, [
             PA0<1>,
             PA5<1>,
             PA15<1>,
@@ -2045,16 +2051,16 @@ pub mod tim2 {
     use crate::timer::ChannelPin;
 
     impl ChannelPin<0> for TIM2 {
-        type Pin = C1;
+        type Pin<Otype> = C1<Otype>;
     }
     impl ChannelPin<1> for TIM2 {
-        type Pin = C2;
+        type Pin<Otype> = C2<Otype>;
     }
     impl ChannelPin<2> for TIM2 {
-        type Pin = C3;
+        type Pin<Otype> = C3<Otype>;
     }
     impl ChannelPin<3> for TIM2 {
-        type Pin = C4;
+        type Pin<Otype> = C4<Otype>;
     }
 }
 
@@ -2063,25 +2069,25 @@ pub mod tim3 {
     use super::*;
 
     pin! {
-        <C1> for [
+        <C1> for PushPull: no:NoPin, [
             PA6<2>,
             PB4<2>,
             PC6<2>,
         ],
-        <C2> for [
+        <C2> for PushPull: no:NoPin, [
             PA7<2>,
             PB5<2>,
             PC7<2>,
         ],
-        <C3> for [
+        <C3> for PushPull: no:NoPin, [
             PB0<2>,
             PC8<2>,
         ],
-        <C4> for [
+        <C4> for PushPull: no:NoPin, [
             PB1<2>,
             PC9<2>,
         ],
-        <Etr> for [
+        <Etr> for PushPull: no:NoPin, [
             PD2<2>,
         ],
     }
@@ -2090,16 +2096,16 @@ pub mod tim3 {
     use crate::timer::ChannelPin;
 
     impl ChannelPin<0> for TIM3 {
-        type Pin = C1;
+        type Pin<Otype> = C1<Otype>;
     }
     impl ChannelPin<1> for TIM3 {
-        type Pin = C2;
+        type Pin<Otype> = C2<Otype>;
     }
     impl ChannelPin<2> for TIM3 {
-        type Pin = C3;
+        type Pin<Otype> = C3<Otype>;
     }
     impl ChannelPin<3> for TIM3 {
-        type Pin = C4;
+        type Pin<Otype> = C4<Otype>;
     }
 }
 
@@ -2108,23 +2114,23 @@ pub mod tim4 {
     use super::*;
 
     pin! {
-        <C1> for [
+        <C1> for PushPull: no:NoPin, [
             PB6<2>,
             PD12<2>,
         ],
-        <C2> for [
+        <C2> for PushPull: no:NoPin, [
             PB7<2>,
             PD13<2>,
         ],
-        <C3> for [
+        <C3> for PushPull: no:NoPin, [
             PB8<2>,
             PD14<2>,
         ],
-        <C4> for [
+        <C4> for PushPull: no:NoPin, [
             PB9<2>,
             PD15<2>,
         ],
-        <Etr> for [
+        <Etr> for PushPull: no:NoPin, [
             PE0<2>,
         ],
     }
@@ -2133,16 +2139,16 @@ pub mod tim4 {
     use crate::timer::ChannelPin;
 
     impl ChannelPin<0> for TIM4 {
-        type Pin = C1;
+        type Pin<Otype> = C1<Otype>;
     }
     impl ChannelPin<1> for TIM4 {
-        type Pin = C2;
+        type Pin<Otype> = C2<Otype>;
     }
     impl ChannelPin<2> for TIM4 {
-        type Pin = C3;
+        type Pin<Otype> = C3<Otype>;
     }
     impl ChannelPin<3> for TIM4 {
-        type Pin = C4;
+        type Pin<Otype> = C4<Otype>;
     }
 }
 
@@ -2150,7 +2156,7 @@ pub mod tim5 {
     use super::*;
 
     pin! {
-        <C1> for [
+        <C1> for PushPull: no:NoPin, [
             PA0<2>,
 
             #[cfg(feature = "gpio-f410")]
@@ -2162,7 +2168,7 @@ pub mod tim5 {
             #[cfg(any(feature = "gpio-f417", feature = "gpio-f427", feature = "gpio-f469"))]
             PH10<2>,
         ],
-        <C2> for [
+        <C2> for PushPull: no:NoPin, [
             PA1<2>,
 
             #[cfg(feature = "gpio-f410")]
@@ -2174,7 +2180,7 @@ pub mod tim5 {
             #[cfg(any(feature = "gpio-f417", feature = "gpio-f427", feature = "gpio-f469"))]
             PH11<2>,
         ],
-        <C3> for [
+        <C3> for PushPull: no:NoPin, [
             PA2<2>,
 
             #[cfg(feature = "gpio-f410")]
@@ -2186,7 +2192,7 @@ pub mod tim5 {
             #[cfg(any(feature = "gpio-f417", feature = "gpio-f427", feature = "gpio-f469"))]
             PH12<2>,
         ],
-        <C4> for [
+        <C4> for PushPull: no:NoPin, [
             PA3<2>,
 
             #[cfg(feature = "gpio-f410")]
@@ -2204,16 +2210,16 @@ pub mod tim5 {
     use crate::timer::ChannelPin;
 
     impl ChannelPin<0> for TIM5 {
-        type Pin = C1;
+        type Pin<Otype> = C1<Otype>;
     }
     impl ChannelPin<1> for TIM5 {
-        type Pin = C2;
+        type Pin<Otype> = C2<Otype>;
     }
     impl ChannelPin<2> for TIM5 {
-        type Pin = C3;
+        type Pin<Otype> = C3<Otype>;
     }
     impl ChannelPin<3> for TIM5 {
-        type Pin = C4;
+        type Pin<Otype> = C4<Otype>;
     }
 }
 
@@ -2229,52 +2235,52 @@ pub mod tim8 {
     use super::*;
 
     pin! {
-        <C1> for [
+        <C1> for PushPull: no:NoPin, [
             PC6<3>,
 
             #[cfg(any(feature = "gpio-f417", feature = "gpio-f427", feature = "gpio-f469"))]
             PI5<3>,
         ],
-        <C2> for [
+        <C2> for PushPull: no:NoPin, [
             PC7<3>,
 
             #[cfg(any(feature = "gpio-f417", feature = "gpio-f427", feature = "gpio-f469"))]
             PI6<3>,
         ],
-        <C3> for [
+        <C3> for PushPull: no:NoPin, [
             PC8<3>,
 
             #[cfg(any(feature = "gpio-f417", feature = "gpio-f427", feature = "gpio-f469"))]
             PI7<3>,
         ],
-        <C4> for [
+        <C4> for PushPull: no:NoPin, [
             PC9<3>,
 
             #[cfg(any(feature = "gpio-f417", feature = "gpio-f427", feature = "gpio-f469"))]
             PI2<3>,
         ],
-        <NC1> for [
+        <NC1> for PushPull: no:NoPin, [
             PA5<3>,
             PA7<3>,
 
             #[cfg(any(feature = "gpio-f417", feature = "gpio-f427", feature = "gpio-f469"))]
             PH13<3>,
         ],
-        <NC2> for [
+        <NC2> for PushPull: no:NoPin, [
             PB0<3>,
             PB14<3>,
 
             #[cfg(any(feature = "gpio-f417", feature = "gpio-f427", feature = "gpio-f469"))]
             PH14<3>,
         ],
-        <NC3> for [
+        <NC3> for PushPull: no:NoPin, [
             PB1<3>,
             PB15<3>,
 
             #[cfg(any(feature = "gpio-f417", feature = "gpio-f427", feature = "gpio-f469"))]
             PH15<3>,
         ],
-        <Etr> for [
+        <Etr> for PushPull: no:NoPin, [
             PA0<3>,
 
             #[cfg(any(feature = "gpio-f412", feature = "gpio-f413"))]
@@ -2283,7 +2289,7 @@ pub mod tim8 {
             #[cfg(any(feature = "gpio-f417", feature = "gpio-f427", feature = "gpio-f469"))]
             PI3<3>,
         ],
-        <BkIn> for [
+        <BkIn> for PushPull: no:NoPin, [
             PA6<3>,
 
             #[cfg(any(feature = "gpio-f412", feature = "gpio-f413"))]
@@ -2298,25 +2304,25 @@ pub mod tim8 {
     use crate::timer::ChannelPin;
 
     impl ChannelPin<0> for TIM8 {
-        type Pin = C1;
+        type Pin<Otype> = C1<Otype>;
     }
     impl ChannelPin<1> for TIM8 {
-        type Pin = C2;
+        type Pin<Otype> = C2<Otype>;
     }
     impl ChannelPin<2> for TIM8 {
-        type Pin = C3;
+        type Pin<Otype> = C3<Otype>;
     }
     impl ChannelPin<3> for TIM8 {
-        type Pin = C4;
+        type Pin<Otype> = C4<Otype>;
     }
     impl ChannelPin<0, true> for TIM8 {
-        type Pin = NC1;
+        type Pin<Otype> = NC1<Otype>;
     }
     impl ChannelPin<1, true> for TIM8 {
-        type Pin = NC2;
+        type Pin<Otype> = NC2<Otype>;
     }
     impl ChannelPin<2, true> for TIM8 {
-        type Pin = NC3;
+        type Pin<Otype> = NC3<Otype>;
     }
 }
 
@@ -2324,7 +2330,7 @@ pub mod tim9 {
     use super::*;
 
     pin! {
-        <C1> for [
+        <C1> for PushPull: no:NoPin, [
             PA2<3>,
 
             #[cfg(not(feature = "gpio-f410"))]
@@ -2333,7 +2339,7 @@ pub mod tim9 {
             #[cfg(feature = "gpio-f410")]
             PC4<3>,
         ],
-        <C2> for [
+        <C2> for PushPull: no:NoPin, [
             PA3<3>,
 
             #[cfg(not(feature = "gpio-f410"))]
@@ -2348,10 +2354,10 @@ pub mod tim9 {
     use crate::timer::ChannelPin;
 
     impl ChannelPin<0> for TIM9 {
-        type Pin = C1;
+        type Pin<Otype> = C1<Otype>;
     }
     impl ChannelPin<1> for TIM9 {
-        type Pin = C2;
+        type Pin<Otype> = C2<Otype>;
     }
 }
 
@@ -2360,7 +2366,7 @@ pub mod tim10 {
     use super::*;
 
     pin! {
-        <C1> for [
+        <C1> for PushPull: no:NoPin, [
             PB8<3>,
 
             #[cfg(any(
@@ -2379,7 +2385,7 @@ pub mod tim10 {
     use crate::timer::ChannelPin;
 
     impl ChannelPin<0> for TIM10 {
-        type Pin = C1;
+        type Pin<Otype> = C1<Otype>;
     }
 }
 
@@ -2387,7 +2393,7 @@ pub mod tim11 {
     use super::*;
 
     pin! {
-        <C1> for [
+        <C1> for PushPull: no:NoPin, [
             PB9<3>,
 
             #[cfg(feature = "gpio-f410")]
@@ -2409,7 +2415,7 @@ pub mod tim11 {
     use crate::timer::ChannelPin;
 
     impl ChannelPin<0> for TIM11 {
-        type Pin = C1;
+        type Pin<Otype> = C1<Otype>;
     }
 }
 
@@ -2425,13 +2431,13 @@ pub mod tim12 {
     use super::*;
 
     pin! {
-        <C1> for [
+        <C1> for PushPull: no:NoPin, [
             PB14<9>,
 
             #[cfg(any(feature = "gpio-f417", feature = "gpio-f427", feature = "gpio-f469"))]
             PH6<9>,
         ],
-        <C2> for [
+        <C2> for PushPull: no:NoPin, [
             PB15<9>,
 
             #[cfg(any(feature = "gpio-f417", feature = "gpio-f427", feature = "gpio-f469"))]
@@ -2443,10 +2449,10 @@ pub mod tim12 {
     use crate::timer::ChannelPin;
 
     impl ChannelPin<0> for TIM12 {
-        type Pin = C1;
+        type Pin<Otype> = C1<Otype>;
     }
     impl ChannelPin<1> for TIM12 {
-        type Pin = C2;
+        type Pin<Otype> = C2<Otype>;
     }
 }
 
@@ -2462,7 +2468,7 @@ pub mod tim13 {
     use super::*;
 
     pin! {
-        <C1> for [
+        <C1> for PushPull: no:NoPin, [
             PA6<9>,
             PF8<9>,
         ],
@@ -2472,7 +2478,7 @@ pub mod tim13 {
     use crate::timer::ChannelPin;
 
     impl ChannelPin<0> for TIM13 {
-        type Pin = C1;
+        type Pin<Otype> = C1<Otype>;
     }
 }
 
@@ -2488,7 +2494,7 @@ pub mod tim14 {
     use super::*;
 
     pin! {
-        <C1> for [
+        <C1> for PushPull: no:NoPin, [
             PA7<9>,
             PF9<9>,
         ],
@@ -2498,6 +2504,6 @@ pub mod tim14 {
     use crate::timer::ChannelPin;
 
     impl ChannelPin<0> for TIM14 {
-        type Pin = C1;
+        type Pin<Otype> = C1<Otype>;
     }
 }