+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
- - Use `enum`s for alternate peripheral pins [#594]
+ - Use `enum`s for alternate peripheral pins (generic over otype) [#594] [#596] [#600]
  - Added missing U(S)ART DMA traits for HAL serial types [#593]
  - Improve SPI::new* docs [#587]
  - Add advanced timer dead time insertion example [#585]
@@ -49,7 +49,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 [#581]: https://github.com/stm32-rs/stm32f4xx-hal/pull/581
 [#594]: https://github.com/stm32-rs/stm32f4xx-hal/pull/594
 [#595]: https://github.com/stm32-rs/stm32f4xx-hal/pull/595
+[#596]: https://github.com/stm32-rs/stm32f4xx-hal/pull/596
 [#599]: https://github.com/stm32-rs/stm32f4xx-hal/pull/599
+[#600]: https://github.com/stm32-rs/stm32f4xx-hal/pull/600
 
 
 ## [v0.14.0] - 2022-12-12